+++ app/constants/collect.js
@@ -2,5 +2,5 @@
 
 module.exports = {
     PLAYER_SPEED: 300,
-    MAPS_TIME_LIMIT: [15, 20, 30]
+    MAPS_TIME_LIMIT: 5
 };