+++ app/constants/versus.js
@@ -1,6 +1,0 @@
-'use strict';
-
-module.exports = {
-    VERSUS_KICKING_POINTS: 5,
-    VERSUS_BOXING_POINTS: 3
-};