+++ app/state-manager.js
@@ -1,18 +1,12 @@
 import BootstrapState from './states/bootstrap-state';
 import LoadingState from './states/loading-state';
-import CollectDragonBallsState from './states/collect-states/collect-dragon-balls-state';
-import EnemyPresentationState from './states/static-states/enemy-presentation-state';
+import CollectDragonBallsState from './states/play-states/collect-dragon-balls-state';
 import GameOverState from './states/static-states/game-over-state';
 import SelectLanguageState from './states/select-states/select-language-state';
-import MealState from './states/static-states/meal-state';
 import SelectPlayerState from './states/select-states/select-player-state';
-import PlayerPresentationState from './states/static-states/player-presentation-state';
-import PlayerHaloPresentationState from './states/static-states/player-halo-presentation-state';
-import ShenronState from './states/static-states/shenron-state';
-import TrainingState from './states/fight-states/training-state';
-import VersusState from './states/fight-states/versus-state';
+import FightState from './states/play-states/fight-state';
 import WinnerState from './states/static-states/winner-state';
-import AvoidMonstersState from './states/fly-states/avoid-monsters-state';
+import AvoidMonstersState from './states/play-states/avoid-monsters-state';
 
 export class StateManager {
     game = null;
@@ -27,19 +21,13 @@ export class StateManager {
     setupStates() {
         this.game.state.add('Bootstrap', BootstrapState);
         this.game.state.add('Loading', LoadingState);
-        this.game.state.add('CollectDragonBalls', CollectDragonBallsState);
-        this.game.state.add('EnemyPresentation', EnemyPresentationState);
-        this.game.state.add('GameOver', GameOverState);
         this.game.state.add('SelectLanguage', SelectLanguageState);
-        this.game.state.add('Meal', MealState);
         this.game.state.add('SelectPlayer', SelectPlayerState);
-        this.game.state.add('PlayerPresentation', PlayerPresentationState);
-        this.game.state.add('PlayerHaloPresentation', PlayerHaloPresentationState);
-        this.game.state.add('Shenron', ShenronState);
-        this.game.state.add('Training', TrainingState);
-        this.game.state.add('Versus', VersusState);
-        this.game.state.add('Winner', WinnerState);
+        this.game.state.add('CollectDragonBalls', CollectDragonBallsState);
         this.game.state.add('AvoidMonsters', AvoidMonstersState);
+        this.game.state.add('Fight', FightState);
+        this.game.state.add('Winner', WinnerState);
+        this.game.state.add('GameOver', GameOverState);
     }
 
     setupNativeListeners() {
@@ -62,61 +50,44 @@ export class StateManager {
             this.game.state.start('SelectLanguage');
         });
 
-        this.game.on('game:language-selected', () => {
+        let selectPlayer = () => {
             this.game.state.start('SelectPlayer');
-        });
+        };
 
-        this.game.on('game:player-halo-presents', () => {
+        this.game.on('game:language-selected', selectPlayer);
+        this.game.on('game:over:try-again', selectPlayer);
+        this.game.on('game:win:try-again', selectPlayer);
+
+        this.game.on('player:select', () => {
             // this.game.state.start('AvoidMonsters');
             this.game.state.start('CollectDragonBalls');
         });
 
-        this.game.on('game:collect-dragon-balls', () => {
-            this.game.state.start('Shenron');
+        let startFight = () => {
+            this.game.state.start('Fight');
+        };
+
+        this.game.on('game:collect-failed', startFight);
+        this.game.on('game:next-enemy', startFight);
+
+        this.game.on('enemy:killed', () => {
+            if (this.game.enemies.length === 0) {
+                this.game.emit('game:win');
+            } else {
+                this.game.emit('game:next-enemy');
+            }
         });
 
-        this.game.on('player:select', () => {
-            this.game.state.start('PlayerHaloPresentation');
-        });
-
-        this.game.on('game:shenron', () => {
-            this.game.state.start('PlayerPresentation');
-        });
-
-        this.game.on('game:player-presents', () => {
-            this.game.state.start('EnemyPresentation');
-        });
-
-        this.game.on('game:training-finished', () => {
-            this.game.state.start('EnemyPresentation');
-        });
-
-        this.game.on('game:enemy-presents', () => {
-            this.game.state.start('Versus');
-        });
-
-        this.game.on('game:player-rest', () => {
-            this.game.state.start('Meal');
-        });
-
-        this.game.on('game:rest-finished', () => {
-            this.game.state.start('Training');
+        this.game.on('game:collect-completed', () => {
+            this.game.state.start('Winner');
         });
 
         this.game.on('game:over', () => {
             this.game.state.start('GameOver');
         });
 
-        this.game.on('game:over:try-again', () => {
-            this.game.state.start('SelectPlayer');
-        });
-
         this.game.on('game:win', () => {
             this.game.state.start('Winner');
-        });
-
-        this.game.on('game:win:try-again', () => {
-            this.game.state.start('SelectPlayer');
         });
     }