+++ app/states/loading-state.js
@@ -1,7 +1,3 @@
-let debug = {
-    log: require('debug')('7balls:loading-state:log')
-};
-
 let utils = require('../helpers/utils');
 let { addSaiyanLabel } = require('../helpers/message');
 let { displayGameVersion } = require('../helpers/message');
@@ -18,45 +14,21 @@ export default class LoadingState extends Phaser.State {
         let progressBar = this.add.sprite(this.world.centerX - (pixelLoadingWidth / 2), 230, 'pixel-loading');
         this.load.setPreloadSprite(progressBar);
 
-        debug.log('assets loading...');
-
-        // graphics/backgrounds
-        this.load.path = './assets/graphics/backgrounds/';
-
-        this.load.image('bg-training-capsule', 'fight/training/bg-training-capsule.png');
-        this.load.image('bg-versus-hell', 'fight/versus/bg-versus-hell.png');
-        this.load.image('bg-versus-heaven', 'fight/versus/bg-versus-heaven.png');
-
-        this.load.image('bg-select-player', 'select/select-player/bg-select-player.png');
-
-        this.load.image('bg-enemy-presentation', 'static/enemy-presentation/bg-enemy-presentation.png');
-        this.load.image('bg-player-presentation', 'static/player-presentation/bg-player-presentation.png');
-        this.load.image('bg-game-over', 'static/game-over/bg-game-over.png');
-        this.load.image('bg-message', 'static/message/bg-message.png');
-        this.load.image('bg-winner', 'static/winner/bg-winner.png');
-        this.load.image('bg-meal-house', 'static/meal/bg-meal-house.png');
-        this.load.image('bg-shenron', 'static/shenron/bg-shenron.png');
-        this.load.image('bg-shenron-growing', 'static/shenron/bg-shenron-growing.png');
-
-        // graphics/bars
-        this.load.path = './assets/graphics/bars/';
-
-        this.load.image('bar-blank', 'blank.png');
-        this.load.image('bar-exp', 'exp.png');
-        this.load.image('bar-exp-invert', 'exp-invert.png');
-        this.load.image('bar-hp', 'hp.png');
-        this.load.image('bar-hp-invert', 'hp-invert.png');
-
-        // graphics/buttons
-        this.load.path = './assets/graphics/buttons/';
-
-        this.load.image('btn-pl', 'pl-flag.png');
-        this.load.image('btn-try-again', 'try-again.png');
-        this.load.image('btn-try-again', 'try-again.png');
-        this.load.image('btn-en', 'usa-flag.png');
-
         // graphics/characters
         this.load.path = './assets/graphics/characters/';
+
+        this.load.image('son-goku-collecting', 'son-goku/son-goku-collecting.png');
+        this.load.image('son-goku', 'son-goku/poster/son-goku.png');
+        this.load.image('son-goku-halo', 'son-goku/poster/son-goku-halo.png');
+        this.load.spritesheet('son-goku-spritesheet', 'son-goku/son-goku-fight.png', 150, 200);
+
+        this.load.image('vegeta-collecting', 'vegeta/vegeta-collecting.png');
+        this.load.image('vegeta', 'vegeta/poster/vegeta.png');
+        this.load.image('vegeta-halo', 'vegeta/poster/vegeta-halo.png');
+        this.load.spritesheet('vegeta-spritesheet', 'vegeta/vegeta-fight.png', 150, 200);
+
+        // graphics/monsters
+        this.load.path = './assets/graphics/monsters/';
 
         this.load.image('bubu', 'bubu/poster/bubu.png');
         this.load.image('bubu-card', 'bubu/bubu-card.png');
@@ -70,47 +42,44 @@ export default class LoadingState extends Phaser.State {
         this.load.image('freeza', 'freeza/poster/freeza.png');
         this.load.spritesheet('freeza-spritesheet', 'freeza/freeza-fight.png', 150, 200);
 
-        this.load.image('son-goku-collecting', 'son-goku/son-goku-collecting.png');
-        this.load.image('son-goku-card', 'son-goku/son-goku-card.png');
-        this.load.image('son-goku', 'son-goku/poster/son-goku.png');
-        this.load.image('son-goku-halo', 'son-goku/poster/son-goku-halo.png');
-        this.load.spritesheet('son-goku-spritesheet', 'son-goku/son-goku-fight.png', 150, 200);
-
         this.load.image('piccolo-collecting', 'piccolo/piccolo-collecting.png');
         this.load.image('piccolo-card', 'piccolo/piccolo-card.png');
         this.load.image('piccolo', 'piccolo/poster/piccolo.png');
         this.load.image('piccolo-halo', 'piccolo/poster/piccolo-halo.png');
         this.load.spritesheet('piccolo-spritesheet', 'piccolo/piccolo-fight.png', 150, 200);
 
-        this.load.image('vegeta-collecting', 'vegeta/vegeta-collecting.png');
-        this.load.image('vegeta-card', 'vegeta/vegeta-card.png');
-        this.load.image('vegeta', 'vegeta/poster/vegeta.png');
-        this.load.image('vegeta-halo', 'vegeta/poster/vegeta-halo.png');
-        this.load.spritesheet('vegeta-spritesheet', 'vegeta/vegeta-fight.png', 150, 200);
-
         this.load.path = '';
-
-        // graphics/icons
-
-        this.load.image('mute', './assets/graphics/icons/mute.png');
-        this.load.image('unmute', './assets/graphics/icons/unmute.png');
-        this.load.image('pixel', './assets/graphics/icons/pixel.png');
-
-        // graphics/logo
-
-        this.load.image('logo-minimal', './assets/graphics/logo/logo-minimal.png');
-
-        debug.log('images files loaded');
 
         this.load.tilemap('collecting-1', './assets/data/maps/collecting-1.json', null, Phaser.Tilemap.TILED_JSON);
         this.load.tilemap('collecting-2', './assets/data/maps/collecting-2.json', null, Phaser.Tilemap.TILED_JSON);
         this.load.tilemap('collecting-3', './assets/data/maps/collecting-3.json', null, Phaser.Tilemap.TILED_JSON);
 
-        debug.log('tilemap files loaded');
+        // ---
+        this.load.path = 'assets/pixelart/';
 
-        this.load.spritesheet('spr-collecting', './assets/graphics/spritesheet/spr-collecting.png', 40, 40);
+        this.load.image('bg-game-over', 'backgrounds/game-over/game-over.png');
+        this.load.image('bg-win', 'backgrounds/win/win.png');
 
-        debug.log('spritesheet files loaded');
+        this.load.image('pixel', 'pixel.png');
+        this.load.image('logo', 'logo/logo.png');
+        this.load.image('flag-pl', 'flags/flag-pl.png');
+        this.load.image('flag-usa', 'flags/flag-usa.png');
+        this.load.image('btn-try-again', 'try-again.png');
+
+        this.load.image('bg-select-player', 'backgrounds/select-player/bg-select-player.png');
+
+        this.load.image('son-goku-card', 'posters/son-goku/son-goku-card.png');
+        this.load.image('vegeta-card', 'posters/vegeta/vegeta-card.png');
+
+        this.load.spritesheet('collect-spritesheet', 'sheets/collecting.png', 40, 40);
+
+        this.load.image('bar-blank', 'bars/blank.png');
+        this.load.image('bar-exp', 'bars/exp.png');
+        this.load.image('bar-exp-invert', 'bars/exp-invert.png');
+        this.load.image('bar-hp', 'bars/hp.png');
+        this.load.image('bar-hp-invert', 'bars/hp-invert.png');
+
+        this.load.path = '';
 
         this.load.json('positions-1', './assets/data/balls-positions/positions-1.json');
         this.load.json('positions-2', './assets/data/balls-positions/positions-2.json');
@@ -119,11 +88,7 @@ export default class LoadingState extends Phaser.State {
         this.load.json('locale-en', './assets/locale/en_EN.json');
         this.load.json('locale-pl', './assets/locale/pl_PL.json');
 
-        debug.log('json files loaded');
-
         this.load.text('font-saiyan-sans', './assets/fonts/Saiyan-Sans.ttf');
-
-        debug.log('fonts files loaded');
     }
 
     create() {