+++ public/assets/locale/definition-types.js
@@ -5,17 +5,10 @@
  * @property {string} CONTROLS_INFO
  * @property {string} COLLECT_DRAGON_BALLS
  *
- * @property {string} COLLECTING_STATE_TIME
- *
- * @property {string} ENEMY_PRESENTATION_STATE_WELCOME
+ * @property {string} TIME
  *
  * @property {string} FIGHT_STATE_LEVEL_SHORT
- *
- * @property {string} MEAL_STATE_WELCOME
- *
- * @property {string} TRAINING_STATE_WELCOME
- *
- * @property {string} VERSUS_STATE_WELCOME
- * @property {string} VERSUS_STATE_PLAYER_DIED
- * @property {string} VERSUS_STATE_PLAYER_WIN
+ * @property {string} FIGHT_STATE_WELCOME
+ * @property {string} FIGHT_STATE_PLAYER_DIED
+ * @property {string} FIGHT_STATE_PLAYER_WIN
  */